# Conflicts:
#	horace_core/sqw/@sqw/private/cut_accumulate_data_.m

+++ horace_core/algorithms/alignment/finalize_alignment.m
@@ -1,5 +1,5 @@
 function [sqw_obj,al_info] = finalize_alignment(filename,varargin)
-%APPLY_ALIGNMEMT Utility takes aligned file and apply alignment information
+%FINALIZE_ALIGNMEMT Utility takes aligned file and apply alignment information
 %stored in it to pixels and image so the object becomes realigned as from
 % the beginning and the aligment information is not necessary any more.
 %
@@ -32,7 +32,7 @@ if ~ld.sqw_type % can not realign dnd object
     return;
 end
 pmd = ld.get_pix_metadata();
-if ~pmd.is_misaligned % nothing to do
+if ~pmd.is_misaligned && pmd.is_range_valid % nothing to do
     al_info = [];
     ld.delete();
     return;
@@ -44,10 +44,8 @@ if keep_original
     argi = {};
 else
     % will replace original
-    argi = {ld.full_filename};    
+    argi = {ld.full_filename};
 end
 [sqw_obj,al_info] = sqw_obj.finalize_alignment(argi{:});
 % TODO: Re #1320 -- add save(obj, file, '-no_update') if called without
 % output arguments
-
-
